Add Reset Module
Unlink SDI from mash toggle
