+++ src/training/frame_counter.rs
@@ -75,6 +75,14 @@ fn tick() {
     }
 }
 
+pub fn reset_all(){
+    unsafe {
+        for (index, _frame) in COUNTERS.iter().enumerate() {
+            full_reset(index);
+        }
+    }
+}
+
 pub fn get_command_flag_cat(
     module_accessor: &mut app::BattleObjectModuleAccessor,
 ) {
